Move webhook related files to the pkg/webhook package
